I added the DAO interfaces and made the classes that implement them as <class-name>Impl
This makes the code cleaner
