+++ JtProject/pom.xml
@@ -15,6 +15,7 @@
 	<name>JtSpringProject</name>
 	<description>Spring project for Java Technology</description>
 	<properties>
+		<mysql-connector-java-version>8.0.33</mysql-connector-java-version>
 		<java.version>11</java.version>
 	</properties>
 	<dependencies>
@@ -55,7 +56,7 @@
 		<dependency>
 			<groupId>mysql</groupId>
 			<artifactId>mysql-connector-java</artifactId>
-			<version>8.0.33</version>
+			<version>${mysql-connector-java-version}</version>
 		</dependency>
