+++ JtProject/src/main/java/com/jtspringproject/JtSpringProject/dao/cartDao.java
@@ -2,37 +2,24 @@ package com.jtspringproject.JtSpringProject.dao;
 import java.util.List;
 
 import com.jtspringproject.JtSpringProject.models.Cart;
-import com.jtspringproject.JtSpringProject.models.Category;
 import org.hibernate.SessionFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 @Repository
-public class cartDao {
-    @Autowired
-    private SessionFactory sessionFactory;
+public interface cartDao {
 
-    public void setSessionFactory(SessionFactory sf){ this.sessionFactory =sf;}
+    public void setSessionFactory(SessionFactory sf);
 
     @Transactional
-    public Cart addCart(Cart cart) {
-        this.sessionFactory.getCurrentSession().save(cart);
-        return cart;
-    }
+    public Cart addCart(Cart cart);
 
     @Transactional
-    public List<Cart> getCarts() {
-        return this.sessionFactory.getCurrentSession().createQuery("from CART").list();
-    }
+    public List<Cart> getCarts();
 
     @Transactional
-    public void updateCart(Cart cart) {
-        this.sessionFactory.getCurrentSession().update(cart);
-    }
+    public void updateCart(Cart cart);
 
     @Transactional
-    public void deleteCart(Cart cart) {
-        this.sessionFactory.getCurrentSession().delete(cart);
-    }
+    public void deleteCart(Cart cart);
 }
