+++ JtProject/src/main/java/com/jtspringproject/JtSpringProject/dao/categoryDao.java
@@ -2,60 +2,30 @@ package com.jtspringproject.JtSpringProject.dao;
 
 import java.util.List;
 
-import org.hibernate.Session;
 import org.hibernate.SessionFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.jtspringproject.JtSpringProject.models.Category;
 
 @Repository
-public class categoryDao {
-	@Autowired
-	private SessionFactory sessionFactory;
+public interface categoryDao {
 
-	public void setSessionFactory(SessionFactory sf) {
-		this.sessionFactory = sf;
-	}
+
+	public void setSessionFactory(SessionFactory sf);
 
 	@Transactional
-	public Category addCategory(String name) {
-		Category category = new Category();
-		category.setName(name);
-		this.sessionFactory.getCurrentSession().saveOrUpdate(category);
-		return category;
-	}
+	public Category addCategory(String name);
 
 	@Transactional
-	public List<Category> getCategories() {
-		return this.sessionFactory.getCurrentSession().createQuery("from CATEGORY").list();
-	}
+	public List<Category> getCategories();
 
 	@Transactional
-	public Boolean deletCategory(int id) {
-
-		Session session = this.sessionFactory.getCurrentSession();
-		Object persistanceInstance = session.load(Category.class, id);
-
-		if (persistanceInstance != null) {
-			session.delete(persistanceInstance);
-			return true;
-		}
-		return false;
-	}
+	public Boolean deletCategory(int id);
 
 	@Transactional
-	public Category updateCategory(int id, String name) {
-		Category category = this.sessionFactory.getCurrentSession().get(Category.class, id);
-		category.setName(name);
-
-		this.sessionFactory.getCurrentSession().update(category);
-		return category;
-	}
+	public Category updateCategory(int id, String name);
 
 	@Transactional
-	public Category getCategory(int id) {
-		return this.sessionFactory.getCurrentSession().get(Category.class,id);
-	}
+	public Category getCategory(int id);
 }
