+++ JtProject/src/main/java/com/jtspringproject/JtSpringProject/dao/impl/cartDaoImpl.java
@@ -1,0 +1,42 @@
+package com.jtspringproject.JtSpringProject.dao.impl;
+import java.util.List;
+
+import com.jtspringproject.JtSpringProject.models.Cart;
+import org.hibernate.SessionFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
+
+@Repository
+public class cartDaoImpl implements com.jtspringproject.JtSpringProject.dao.cartDao {
+    @Autowired
+    private SessionFactory sessionFactory;
+
+    public void setSessionFactory(SessionFactory sf){ this.sessionFactory =sf;}
+    
+    @Override
+    @Transactional
+    public Cart addCart(Cart cart) {
+        this.sessionFactory.getCurrentSession().save(cart);
+        return cart;
+    }
+    
+    @Override
+    @SuppressWarnings("unchecked")
+	@Transactional
+    public List<Cart> getCarts() {
+        return this.sessionFactory.getCurrentSession().createQuery("from CART").list();
+    }
+    
+    @Override
+    @Transactional
+    public void updateCart(Cart cart) {
+        this.sessionFactory.getCurrentSession().update(cart);
+    }
+    
+    @Override
+    @Transactional
+    public void deleteCart(Cart cart) {
+        this.sessionFactory.getCurrentSession().delete(cart);
+    }
+}
