+++ JtProject/src/main/java/com/jtspringproject/JtSpringProject/dao/impl/productDaoImpl.java
@@ -1,0 +1,62 @@
+package com.jtspringproject.JtSpringProject.dao.impl;
+
+import java.util.List;
+
+import org.hibernate.Session;
+import org.hibernate.SessionFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.jtspringproject.JtSpringProject.models.Product;
+
+@Repository
+public class productDaoImpl implements com.jtspringproject.JtSpringProject.dao.productDao{
+	@Autowired
+    private SessionFactory sessionFactory;
+	
+	public void setSessionFactory(SessionFactory sf) {
+        this.sessionFactory = sf;
+    }
+	
+	@SuppressWarnings("unchecked")
+	@Override
+	@Transactional
+	public List<Product> getProducts(){
+		return this.sessionFactory.getCurrentSession().createQuery("from PRODUCT").list();
+	}
+	
+	@Override
+	@Transactional
+	public Product addProduct(Product product) {
+		this.sessionFactory.getCurrentSession().save(product);
+		return product;
+	}
+	
+	@Override
+	@Transactional
+	public Product getProduct(int id) {
+		return this.sessionFactory.getCurrentSession().get(Product.class, id);
+	}
+	
+	@Override
+	public Product updateProduct(Product product){
+		this.sessionFactory.getCurrentSession().update(String.valueOf(Product.class),product);
+		return product;
+	}
+	
+	@Override
+	@Transactional
+	public Boolean deletProduct(int id) {
+
+		Session session = this.sessionFactory.getCurrentSession();
+		Object persistanceInstance = session.load(Product.class, id);
+
+		if (persistanceInstance != null) {
+			session.delete(persistanceInstance);
+			return true;
+		}
+		return false;
+	}
+
+}
