+++ JtProject/src/main/java/com/jtspringproject/JtSpringProject/dao/impl/userDaoImpl.java
@@ -1,0 +1,77 @@
+package com.jtspringproject.JtSpringProject.dao.impl;
+
+import java.util.List;
+
+import org.hibernate.Session;
+import org.hibernate.SessionFactory;
+import org.hibernate.query.Query;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.jtspringproject.JtSpringProject.dao.userDao;
+import com.jtspringproject.JtSpringProject.models.User;
+
+
+@Repository
+public class userDaoImpl implements userDao {
+	@Autowired
+    private SessionFactory sessionFactory;
+	
+	public void setSessionFactory(SessionFactory sf) {
+        this.sessionFactory = sf;
+    }
+	
+	@Override
+   @Transactional
+    public List<User> getAllUser() {
+        Session session = this.sessionFactory.getCurrentSession();
+		@SuppressWarnings("unchecked")
+		List<User>  userList = session.createQuery("from CUSTOMER").list();
+        return userList;
+    }
+    
+	@Override
+    @Transactional
+	public User saveUser(User user) {
+		this.sessionFactory.getCurrentSession().saveOrUpdate(user);
+		System.out.println("User added" + user.getId());
+        return user;
+	}
+    
+//    public User checkLogin() {
+//    	this.sessionFactory.getCurrentSession().
+//    }
+	
+	@Override
+    @Transactional
+    public User getUser(String username,String password) {
+    	@SuppressWarnings("rawtypes")
+		Query query = sessionFactory.getCurrentSession().createQuery("from CUSTOMER where username = :username");
+    	query.setParameter("username",username);
+    	
+    	try {
+			User user = (User) query.getSingleResult();
+			System.out.println(user.getPassword());
+			if(password.equals(user.getPassword())) {
+				return user;
+			}else {		
+				return new User();
+			}
+		}catch(Exception e){
+			System.out.println(e.getMessage());
+			User user = new User();
+			return user;
+		}
+    	
+    }
+	
+	@Override
+	@Transactional
+	public boolean userExists(String username) {
+		@SuppressWarnings("rawtypes")
+		Query query = sessionFactory.getCurrentSession().createQuery("from CUSTOMER where username = :username");
+		query.setParameter("username",username);
+		return !query.getResultList().isEmpty();
+	}
+}
