+++ JtProject/src/main/java/com/jtspringproject/JtSpringProject/dao/productDao.java
@@ -2,55 +2,28 @@ package com.jtspringproject.JtSpringProject.dao;
 
 import java.util.List;
 
-import org.hibernate.Session;
 import org.hibernate.SessionFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.jtspringproject.JtSpringProject.models.Category;
 import com.jtspringproject.JtSpringProject.models.Product;
 
 @Repository
-public class productDao {
-	@Autowired
-    private SessionFactory sessionFactory;
+public interface productDao {
 	
-	public void setSessionFactory(SessionFactory sf) {
-        this.sessionFactory = sf;
-    }
+	public void setSessionFactory(SessionFactory sf);
 	
 	@Transactional
-	public List<Product> getProducts(){
-		return this.sessionFactory.getCurrentSession().createQuery("from PRODUCT").list();
-	}
+	public List<Product> getProducts();
 	
 	@Transactional
-	public Product addProduct(Product product) {
-		this.sessionFactory.getCurrentSession().save(product);
-		return product;
-	}
+	public Product addProduct(Product product);
 	
 	@Transactional
-	public Product getProduct(int id) {
-		return this.sessionFactory.getCurrentSession().get(Product.class, id);
-	}
+	public Product getProduct(int id);
 
-	public Product updateProduct(Product product){
-		this.sessionFactory.getCurrentSession().update(String.valueOf(Product.class),product);
-		return product;
-	}
+	public Product updateProduct(Product product);
+	
 	@Transactional
-	public Boolean deletProduct(int id) {
-
-		Session session = this.sessionFactory.getCurrentSession();
-		Object persistanceInstance = session.load(Product.class, id);
-
-		if (persistanceInstance != null) {
-			session.delete(persistanceInstance);
-			return true;
-		}
-		return false;
-	}
-
+	public Boolean deletProduct(int id);
 }
