+++ JtProject/src/main/java/com/jtspringproject/JtSpringProject/dao/userDao.java
@@ -2,13 +2,7 @@ package com.jtspringproject.JtSpringProject.dao;
 
 import java.util.List;
 
-import javax.persistence.NoResultException;
-import javax.sound.midi.Soundbank;
-
-import org.hibernate.Session;
 import org.hibernate.SessionFactory;
-import org.hibernate.query.Query;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,55 +10,21 @@ import com.jtspringproject.JtSpringProject.models.User;
 
 
 @Repository
-public class userDao {
-	@Autowired
-    private SessionFactory sessionFactory;
+public interface userDao {
 	
-	public void setSessionFactory(SessionFactory sf) {
-        this.sessionFactory = sf;
-    }
+	public void setSessionFactory(SessionFactory sf);
    @Transactional
-    public List<User> getAllUser() {
-        Session session = this.sessionFactory.getCurrentSession();
-		List<User>  userList = session.createQuery("from CUSTOMER").list();
-        return userList;
-    }
+    public List<User> getAllUser();
     
     @Transactional
-	public User saveUser(User user) {
-		this.sessionFactory.getCurrentSession().saveOrUpdate(user);
-		System.out.println("User added" + user.getId());
-        return user;
-	}
+	public User saveUser(User user);
     
 //    public User checkLogin() {
 //    	this.sessionFactory.getCurrentSession().
 //    }
     @Transactional
-    public User getUser(String username,String password) {
-    	Query query = sessionFactory.getCurrentSession().createQuery("from CUSTOMER where username = :username");
-    	query.setParameter("username",username);
-    	
-    	try {
-			User user = (User) query.getSingleResult();
-			System.out.println(user.getPassword());
-			if(password.equals(user.getPassword())) {
-				return user;
-			}else {		
-				return new User();
-			}
-		}catch(Exception e){
-			System.out.println(e.getMessage());
-			User user = new User();
-			return user;
-		}
-    	
-    }
+    public User getUser(String username,String password);
 
 	@Transactional
-	public boolean userExists(String username) {
-		Query query = sessionFactory.getCurrentSession().createQuery("from CUSTOMER where username = :username");
-		query.setParameter("username",username);
-		return !query.getResultList().isEmpty();
-	}
+	public boolean userExists(String username);
 }
